Move architecture tests to a dedicated package
