+++ .github/workflows/quality-monitor-jenkins.yml
@@ -1,0 +1,175 @@
+name: 'Quality Monitor PR'
+
+on:
+  pull_request_target:
+
+jobs:
+  build:
+
+    runs-on: [ubuntu-latest]
+    name: Build, test and monitor quality on Ubuntu
+
+    steps:
+      - name: 'Checkout merge commit'
+        uses: actions/checkout@v4
+        with:
+          ref: "${{ github.event.pull_request.merge_commit_sha }}"
+        if: github.event.pull_request.merge_commit_sha != ''
+      - name: 'Checkout PR head commit'
+        uses: actions/checkout@v4
+        with:
+          ref: "${{ github.event.pull_request.head.sha }}"
+        if: github.event.pull_request.merge_commit_sha == ''
+      - name: Set up JDK 21
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'temurin'
+          java-version: 21
+          check-latest: true
+          cache: 'maven'
+      - name: Set up Maven
+        uses: stCarolas/setup-maven@v5
+        with:
+          maven-version: 3.9.9
+      - name: Cache the NVD database
+        uses: actions/cache@v4
+        with:
+          path: ~/.m2/repository/org/owasp/dependency-check-data
+          key: dependency-check
+      - name: Build with Maven
+        env:
+          BROWSER: chrome-container
+          NVD_API_KEY: ${{ secrets.NVD_API_KEY }}
+        run: mvn -V --color always -ntp clean verify -Pci -Powasp | tee maven.log
+      - name: Extract pull request number
+        uses: jwalton/gh-find-current-pr@v1
+        id: pr
+      - name: Run Quality Monitor
+        uses: uhafner/quality-monitor@v2
+        with:
+          pr-number: ${{ steps.pr.outputs.number }}
+          show-headers: true
+          config: >
+            {
+              "tests": {
+                "name": "Tests",
+                "tools": [
+                  {
+                    "id": "junit",
+                    "name": "Unit Tests",
+                    "pattern": "**/target/surefire-reports/TEST*analysis*.xml"
+                  },
+                  {
+                    "id": "junit",
+                    "icon": "rocket",
+                    "name": "Integration tests",
+                    "pattern": "**/target/failsafe-reports/TEST*.xml"
+                  },
+                  {
+                    "id": "junit",
+                    "icon": "no_entry",
+                    "name": "Architecture tests",
+                    "pattern": "**/target/surefire-reports/TEST*archunit*.xml"
+                  }
+                ]
+              },
+              "analysis": [
+                {
+                  "name": "Style",
+                  "id": "style",
+                  "tools": [
+                    {
+                      "id": "checkstyle",
+                      "pattern": "**/target/checkstyle-*/checkstyle-result.xml"
+                    },
+                    {
+                      "id": "pmd",
+                      "pattern": "**/target/pmd-*/pmd.xml"
+                    }
+                  ]
+                },
+                {
+                  "name": "Bugs",
+                  "id": "bugs",
+                  "icon": "bug",
+                  "tools": [
+                    {
+                      "id": "spotbugs",
+                      "sourcePath": "src/main/java",
+                      "pattern": "**/target/spotbugsXml.xml"
+                    }
+                  ]
+                },
+                {
+                  "name": "Vulnerabilities",
+                  "id": "vulnerabilities",
+                  "icon": "shield",
+                  "tools": [
+                    {
+                      "id": "owasp-dependency-check",
+                      "pattern": "**/target/dependency-check-report.json"                    
+                    }
+                  ]
+                }
+              ],
+              "coverage": [
+                {
+                  "name": "Code Coverage",
+                  "tools": [
+                    {
+                      "id": "jacoco",
+                      "metric": "line",
+                      "sourcePath": "src/main/java",
+                      "pattern": "**/target/site/jacoco/jacoco.xml"
+                    },
+                    {
+                      "id": "jacoco",
+                      "metric": "branch",
+                      "sourcePath": "src/main/java",
+                      "pattern": "**/target/site/jacoco/jacoco.xml"
+                    }
+                  ]
+                }
+              ],
+              "metrics": 
+                {
+                  "name": "Software Metrics",
+                  "tools": [
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "CYCLOMATIC_COMPLEXITY"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "COGNITIVE_COMPLEXITY"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "NPATH_COMPLEXITY"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "LOC"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "NCSS"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "COHESION"
+                    },
+                    {
+                      "id": "metrics",
+                      "pattern": "**/metrics/pmd.xml",
+                      "metric": "WEIGHT_OF_CLASS"
+                    }
+                  ]
+                }
+            }
