+++ plugin/src/test/java/io/jenkins/plugins/archunit/PluginArchitectureTest.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins.analysis.warnings;
+package io.jenkins.plugins.archunit;
 
 import com.tngtech.archunit.base.DescribedPredicate;
 import com.tngtech.archunit.core.domain.AccessTarget.ConstructorCallTarget;
@@ -26,6 +26,10 @@ import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.*;
 @SuppressWarnings("hideutilityclassconstructor")
 @AnalyzeClasses(packages = "io.jenkins.plugins.analysis")
 final class PluginArchitectureTest {
+    private PluginArchitectureTest() {
+        // prevents instantiation
+    }
+
     @ArchTest
     static final ArchRule NO_EXCEPTIONS_WITH_NO_ARG_CONSTRUCTOR = noClasses()
             .that().haveSimpleNameNotContaining("Benchmark")
@@ -94,8 +98,5 @@ final class PluginArchitectureTest {
         private boolean isPermittedException(final JavaClass owner) {
             return allowedExceptions.stream().anyMatch(owner::isAssignableTo);
         }
-    }
-
-    private PluginArchitectureTest() {
     }
 }
